Fix t\complete_program.t failure on Windows
On Windows, whether a file is "executable" is determined by file
extension -- chmod 0755 does not do it. A zero byte .BAT file is
a valid executable. On Unixy systems, all that matters are the
permissions bits, so my fix was to change the names of the files
created by the test script to `prog%d.bat`.

See also

http://blog.nu42.com/2014/12/when-bits-dont-stick.html#complete-util-bug